Repackage Jackson Serializers/Deserializers and add constants (#167)
* Constants and packaking


* Please spotsbugs


* Rename Class and move package


* spotless


* make it compile
