+++ flytekit-examples/src/main/java/org/flyte/examples/SimpleStructTask.java
@@ -54,4 +54,3 @@ public class SimpleStructTask
     return Output.create(input.in());
   }
 }
-*/
