+++ flytekit-jackson/src/main/java/org/flyte/flytekit/jackson/JacksonSdkType.java
@@ -38,6 +38,8 @@ import org.flyte.api.v1.LiteralType;
 import org.flyte.api.v1.Variable;
 import org.flyte.flytekit.SdkBindingData;
 import org.flyte.flytekit.SdkType;
+import org.flyte.flytekit.jackson.deserializers.CustomSdkBindingDataDeserializers;
+import org.flyte.flytekit.jackson.deserializers.LiteralMapDeserializer;
 
 public class JacksonSdkType<T> extends SdkType<T> {
