+++ flytekit-jackson/src/main/java/org/flyte/flytekit/jackson/deserializers/SdkBindingDataDeserializers.java
@@ -14,23 +14,25 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.flyte.flytekit.jackson;
+package org.flyte.flytekit.jackson.deserializers;
 
 import com.fasterxml.jackson.databind.BeanDescription;
+import com.fasterxml.jackson.databind.DeserializationConfig;
 import com.fasterxml.jackson.databind.JavaType;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializationConfig;
-import com.fasterxml.jackson.databind.ser.Serializers;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import com.fasterxml.jackson.databind.deser.Deserializers;
+import org.flyte.flytekit.SdkBindingData;
 
-public class LiteralMapSerializers extends Serializers.Base {
+public class SdkBindingDataDeserializers extends Deserializers.Base {
 
   @Override
-  public JsonSerializer<?> findSerializer(
-      SerializationConfig config, JavaType type, BeanDescription beanDesc) {
-    if (type.getRawClass().equals(JacksonLiteralMap.class)) {
-      return new LiteralMapSerializer();
+  public JsonDeserializer<?> findBeanDeserializer(
+      JavaType type, DeserializationConfig config, BeanDescription beanDesc) {
+    if (SdkBindingData.class.isAssignableFrom(type.getRawClass())) {
+
+      return new SdkBindingDataDeserializer();
     }
 
-    return super.findSerializer(config, type, beanDesc);
+    return null;
   }
 }
