+++ flytekit-jackson/src/main/java/org/flyte/flytekit/jackson/serializers/BooleanSerializer.java
@@ -14,7 +14,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.flyte.flytekit.jackson;
+package org.flyte.flytekit.jackson.serializers;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
@@ -22,36 +22,26 @@ import java.io.IOException;
 import org.flyte.api.v1.Literal;
 import org.flyte.api.v1.LiteralType;
 import org.flyte.api.v1.Primitive;
-import org.flyte.api.v1.Scalar;
+import org.flyte.api.v1.SimpleType;
 
-public abstract class PrimitiveSerializer extends ScalarSerializer {
-
-  public PrimitiveSerializer(
+public class BooleanSerializer extends PrimitiveSerializer {
+  public BooleanSerializer(
       JsonGenerator gen,
       String key,
       Literal value,
       SerializerProvider serializerProvider,
       LiteralType literalType) {
     super(gen, key, value, serializerProvider, literalType);
+    if (literalType.getKind() != LiteralType.Kind.SIMPLE_TYPE
+        && literalType.simpleType() != SimpleType.BOOLEAN) {
+      throw new IllegalArgumentException("Literal type should be a boolean literal type");
+    }
   }
 
   @Override
-  public final void serializeScalar() throws IOException {
-    gen.writeObject(Scalar.Kind.PRIMITIVE);
-    serializePrimitive();
-  }
-
-  abstract void serializePrimitive() throws IOException;
-
-  protected void writePrimitive(Object kind, WritePrimitiveFunction writeValueFunction)
-      throws IOException {
-    gen.writeFieldName("primitive");
-    gen.writeObject(kind);
-    gen.writeFieldName("value");
-    writeValueFunction.write(gen, value.scalar().primitive());
-  }
-
-  interface WritePrimitiveFunction {
-    void write(JsonGenerator gen, Primitive value) throws IOException;
+  public void serializePrimitive() throws IOException {
+    writePrimitive(
+        Primitive.Kind.BOOLEAN_VALUE,
+        (gen, primitive) -> gen.writeBoolean(primitive.booleanValue()));
   }
 }
