+++ flytekit-local-engine/src/test/java/org/flyte/localengine/examples/StructTask.java
@@ -17,6 +17,7 @@
 package org.flyte.localengine.examples;
 
 import com.google.auto.value.AutoValue;
+import org.flyte.flytekit.SdkBindingData;
 import org.flyte.flytekit.SdkRunnableTask;
 import org.flyte.flytekit.jackson.JacksonSdkType;
 
@@ -30,9 +31,11 @@ public class StructTask extends SdkRunnableTask<StructTask.Input, StructTask.Out
 
   @Override
   public Output run(Input input) {
-    return Output.create(
-        CustomStruct.create(
-            input.structData().get().someKey1() + "-output", input.structData().get().someKey2()));
+    return null;
+    //    return Output.create(
+    //        CustomStruct.create(
+    //            input.structData().get().someKey1() + "-output",
+    // input.structData().get().someKey2()));
   }
 
   @AutoValue
@@ -54,7 +57,9 @@ public class StructTask extends SdkRunnableTask<StructTask.Input, StructTask.Out
     public abstract SdkBindingData<CustomStruct> structData();
 
     public static Input create(String someString, CustomStruct struct) {
-      return new AutoValue_StructTask_Input(SdkBindingData.ofString(someString), SdkBindingData.ofStruct(struct));
+      return null;
+      // return new AutoValue_StructTask_Input(SdkBindingData.ofString(someString),
+      // SdkBindingData.ofStruct(struct));
     }
   }
 
@@ -68,4 +73,3 @@ public class StructTask extends SdkRunnableTask<StructTask.Input, StructTask.Out
     }
   }
 }
-*/
