+++ flytekit-testing/src/test/java/org/flyte/flytekit/testing/SdkTestingExecutorTest.java
@@ -454,7 +454,6 @@ public class SdkTestingExecutorTest {
 
     public static SimpleSubWorkflowInput create(SdkBindingData<Long> in) {
       return new AutoValue_SdkTestingExecutorTest_SimpleSubWorkflowInput(in);
-
     }
   }
